Remove obsolete ECS components and refactor game structure
- Deleted unused files related to player movement, star field, and scene management, streamlining the codebase.
- Refactored the game structure to utilize core ECS components, enhancing maintainability and clarity.
- Updated references in the game logic to align with the new component organization, ensuring consistent functionality across scenes and systems.
