+++ internal/ecs/core/factory.go
@@ -1,4 +1,4 @@
-package ecs
+package core
 
 import (
 	"time"
